dacodes/requirements.txt
Django==3.1
djangorestframework==3.11.1
drf-yasg==1.17.1
drf-nested-routers==0.91
django-extensions==3.0.4